requirements.txt
requests>=2.31.0
urllib3>=2.0.0
idna>=3.0
charset-normalizer>=3.0
certifi>=2023.0.0
PyYAML>=6.0
